+++ examples/two_file_example.py
@@ -1,0 +1,76 @@
+from pyshacl import validate
+
+
+shapes_file = '''
+@prefix dash: <http://datashapes.org/dash#> .
+@prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+@prefix schema: <http://schema.org/> .
+@prefix sh: <http://www.w3.org/ns/shacl#> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+schema:PersonShape
+    a sh:NodeShape ;
+    sh:targetClass schema:Person ;
+    sh:property [
+        sh:path schema:givenName ;
+        sh:datatype xsd:string ;
+        sh:name "given name" ;
+    ] ;
+    sh:property [
+        sh:path schema:birthDate ;
+        sh:lessThan schema:deathDate ;
+        sh:maxCount 1 ;
+    ] ;
+    sh:property [
+        sh:path schema:gender ;
+        sh:in ( "female" "male" ) ;
+    ] ;
+    sh:property [
+        sh:path schema:address ;
+        sh:node schema:AddressShape ;
+    ] .
+
+schema:AddressShape
+    a sh:NodeShape ;
+    sh:closed true ;
+    sh:property [
+        sh:path schema:streetAddress ;
+        sh:datatype xsd:string ;
+    ] ;
+    sh:property [
+        sh:path schema:postalCode ;
+        sh:or ( [ sh:datatype xsd:string ] [ sh:datatype xsd:integer ] ) ;
+        sh:minInclusive 10000 ;
+        sh:maxInclusive 99999 ;
+    ] .
+'''
+shapes_file_format = 'turtle'
+
+data_file = '''
+{
+    "@context": { "@vocab": "http://schema.org/" },
+
+    "@id": "http://example.org/ns#Bob",
+    "@type": "Person",
+    "givenName": "Robert",
+    "familyName": "Junior",
+    "birthDate": "1971-07-07",
+    "deathDate": "1968-09-10",
+    "address": {
+        "@id": "http://example.org/ns#BobsAddress",
+        "streetAddress": "1600 Amphitheatre Pkway",
+        "postalCode": 9404
+    }
+}
+'''
+data_file_format = 'json-ld'
+
+conforms, v_graph, v_text = validate(data_file, shacl_graph=shapes_file,
+                                     target_graph_format=data_file_format,
+                                     shacl_graph_format=shapes_file_format,
+                                     inference='rdfs', debug=True,
+                                     serialize_report_graph=True)
+print(conforms)
+print(v_graph)
+print(v_text)